+++ .gitignore
@@ -3,3 +3,5 @@ models
 __pycache__
 .env
 cropped_faces
+.venv
+tflite_models
