+++ Dockerfile
@@ -4,14 +4,6 @@ COPY . /app
 RUN apt-get update && apt-get install -y \
     libgl1-mesa-glx \
     libglib2.0-0 \
-    libsm6 \
-    libxext6 \
-    libxrender-dev \
-    gcc \
-    g++ \
-    libhdf5-dev \
-    libhdf5-serial-dev \
-    hdf5-tools \
     && rm -rf /var/lib/apt/lists/*
 RUN pip install --upgrade pip
 RUN pip install --no-cache-dir -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cpu
