+++ src/services/classify_single_image_service.py
@@ -1,0 +1,74 @@
+from PIL import Image
+from fastapi import HTTPException
+from typing import Dict, Any
+import logging, time
+
+from optimum.pipelines import pipeline
+
+from src.shared.crop_face_from_image import crop_face_from_image
+from src.shared.detect_faces_in_image import detect_faces_in_image
+
+logger = logging.getLogger(__name__)
+
+
+async def process_single_image(
+    img: Image.Image,
+    model: pipeline,
+    detect_faces: bool = False,
+    save_cropped: bool = False,
+) -> Dict[str, Any]:
+    try:
+        start_time = time.time()
+
+        if detect_faces:
+            faces_detected, face_count, face_locations = detect_faces_in_image(img)
+
+            if not faces_detected:
+                end_time = time.time()
+                processing_time = end_time - start_time
+
+                return {
+                    "type": "single_image",
+                    "faces_detected": False,
+                    "face_count": 0,
+                    "predictions": None,
+                    "face_locations": [],
+                    "processing_time": processing_time,
+                }
+            predictions_cropped = []
+            for face_location in face_locations:
+                cropped_face = crop_face_from_image(
+                    img,
+                    face_location,
+                    save_cropped=save_cropped,
+                )
+                prediction_cropped = model(cropped_face)
+                predictions_cropped.append(prediction_cropped)
+
+            end_time = time.time()
+            processing_time = end_time - start_time
+
+            return {
+                "type": "multi_face",
+                "faces_detected": True,
+                "face_count": face_count,
+                "face_locations": face_locations,
+                "predictions": predictions_cropped,
+                "processing_time": processing_time,
+            }
+        else:
+            predictions = model(img)
+
+            end_time = time.time()
+            processing_time = end_time - start_time
+
+            return {
+                "type": "single_image",
+                "faces_detected": False,
+                "face_count": 0,
+                "predictions": predictions,
+                "processing_time": processing_time,
+            }
+    except Exception as e:
+        logger.error(f"Error processing single image: {e}", exc_info=True)
+        raise HTTPException(status_code=500, detail="Error processing image")
